+++ CHANGELOG.md
@@ -30,6 +30,7 @@ implements: [#46](https://github.com/pdir/contao-survey/issues/46) 🤗 [AI ANTW
 - [Fixed] remove MySqlPlatform usage [#40](https://github.com/pdir/contao-survey/issues/40)
 - [Fixed] Fix immediate_start check [PR](https://github.com/pdir/contao-survey/pull/69) 🤗 [fritzmg](https://github.com/fritzmg)
 - [Fixed] Do not use generateFrontendUrl [PR](https://github.com/pdir/contao-survey/pull/67) 🤗 [fritzmg](https://github.com/fritzmg)
+- [Fixed] Fix error in survey element [#66](https://github.com/pdir/contao-survey/issues/66) 🤗 [arminfrey](https://github.com/arminfrey)
 
 ## [3.5.1](https://github.com/pdir/contao-survey/tree/3.5.1) – 2023-08-20